% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_templates.R
\name{use_codespaces}
\alias{use_codespaces}
\title{add codespaces config files for using r
This sets up vs code with helpful packages and extensions for running r code,
including radian.}
\usage{
use_codespaces()
}
\description{
add codespaces config files for using r
This sets up vs code with helpful packages and extensions for running r code,
including radian.
}
